feat: dataset description full width
