+++ tests/component/sections/collection/collection-items-panel/dataset-card/DatasetCard.spec.tsx
@@ -1,6 +1,7 @@
 import { DatasetCard } from '@/sections/collection/collection-items-panel/items-list/dataset-card/DatasetCard'
 import { DatasetItemTypePreviewMother } from '@tests/component/dataset/domain/models/DatasetItemTypePreviewMother'
 import { DateHelper } from '@/shared/helpers/DateHelper'
+import styles from '@/sections/collection/collection-items-panel/items-list/dataset-card/DatasetCard.module.scss'
 
 describe('DatasetCard', () => {
   it('should render the card', () => {
@@ -14,6 +15,7 @@ describe('DatasetCard', () => {
     cy.findByText(DateHelper.toDisplayFormat(dataset.releaseOrCreateDate)).should('exist')
     cy.findByText(/Admin, Dataverse, 2023, "Dataset Title",/).should('exist')
   })
+
   it('should render the card with user roles', () => {
     const userRoles = ['Admin', 'Contributor']
     const dataset = DatasetItemTypePreviewMother.create({ userRoles: userRoles })
@@ -28,5 +30,29 @@ describe('DatasetCard', () => {
     userRoles.forEach((role) => {
       cy.findByText(role).should('exist')
     })
+  })
+
+  it('should render the dataset info correctly', () => {
+    const dataset = DatasetItemTypePreviewMother.createDraft()
+    cy.customMount(<DatasetCard datasetPreview={dataset} />)
+
+    cy.findByText(DateHelper.toDisplayFormat(dataset.releaseOrCreateDate)).should('exist')
+    cy.findByText(/Admin, Dataverse, 2023, "Dataset Title",/)
+      .should('exist')
+      .parent()
+      .parent()
+      .should('have.class', styles['citation-box'])
+    cy.findByText(dataset.description).should('exist')
+  })
+
+  it('should render the citation with the deaccessioned background if the dataset is deaccessioned', () => {
+    const dataset = DatasetItemTypePreviewMother.createDeaccessioned()
+    cy.customMount(<DatasetCard datasetPreview={dataset} />)
+
+    cy.findByText(/Admin, Dataverse, 2023, "Dataset Title",/)
+      .should('exist')
+      .parent()
+      .parent()
+      .should('have.class', styles['deaccesioned'])
   })
 })
